requirements.txt
streamlit
google-generativeai
PyPDF2
reportlab
streamlit-pdf-viewer